full search POST router function using unirest working in postman
