+++ my-express-app/package.json
@@ -14,9 +14,10 @@
     "express": "~4.18.2",
     "morgan": "~1.9.1",
     "mysql": "^2.18.1",
-    "node-fetch": "^3.3.1",
+    "node-fetch": "^2.6.9",
     "nodemon": "^2.0.21",
     "podcast-api": "^1.1.4",
-    "react-router-dom": "^6.9.0"
+    "react-router-dom": "^6.9.0",
+    "unirest": "^0.6.0"
   }
 }
