+++ my-express-app/routes/podcasts.js
@@ -1,45 +1,23 @@
 const express = require("express");
 const router = express.Router();
-const podcastApi = require("podcast-api");
+const { Client } = require("podcast-api");
 
-// Set up authentication
-podcastApi.setAuth({ token: process.env.LISTEN_API_KEY || null });
+const client = Client({
+  apiKey: process.env.LISTEN_API_KEY || null,
+});
 
-// Define a route for searching podcasts
+// FULL SEARCH -- RETURNING SAMPLE DATA -- NOT USING API KEY
 router.get("/search", async (req, res) => {
+  const searchTerm = req.params;
   try {
-    const searchTerm = req.query.term;
-    const limit = req.query.limit || 10;
-    const results = await fetch(podcastApi.search({ term: searchTerm, limit }));
-    res.json(results);
+    const results = await client.search({
+      q: { searchTerm },
+    });
+    // console.log(results);
+    res.status(200).send(results.data.results); // Returns an array of podcasts
+    // console.log(data);
   } catch (error) {
-    console.error(error);
-    res.status(500).send("Server error");
-  }
-});
-
-// Define a route for getting podcast metadata
-router.get("/podcasts/:id", async (req, res) => {
-  try {
-    const id = req.params.id;
-    const podcast = await podcastApi.lookup({ id });
-    res.json(podcast);
-  } catch (error) {
-    console.error(error);
-    res.status(500).send("Server error");
-  }
-});
-
-// Define a route for getting recent podcast episodes
-router.get("/podcasts/:id/episodes", async (req, res) => {
-  try {
-    const id = req.params.id;
-    const limit = req.query.limit || 10;
-    const episodes = await podcastApi.episodes({ id, limit });
-    res.json(episodes);
-  } catch (error) {
-    console.error(error);
-    res.status(500).send("Server error");
+    console.log(error);
   }
 });
